requirements.txt
fastapi>=0.115.6,<0.120.0
uvicorn[standard]>=0.34.0,<0.40.0
langgraph>=0.3.25,<0.7.0
langchain-core>=0.3.60,<1.0.0
langchain-google-genai>=2.0.7,<3.0.0
google-generativeai>=0.8.3,<1.0.0
ag-ui-protocol>=0.1.7,<0.2.0
ag-ui-langgraph>=0.0.9,<0.1.0
tavily-python>=0.5.0,<1.0.0
mem0ai>=0.1.40,<1.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.11.2,<3.0.0
httpx>=0.28.1,<1.0.0
truststore>=0.10.0,<1.0.0

